requirements.txt
aiofiles==24.1.0
anyio==4.4.0
certifi==2024.7.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
m3u8==5.1.0
nest-asyncio==1.6.0
orjson==3.10.6
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
twitter-api-client==0.10.22
uvloop==0.19.0
websockets==12.0
